+++ src/LSST/M1M3/SS/States/EnabledActiveState.h
@@ -39,8 +39,6 @@ public:
     States::Type lowerM1M3(LowerM1M3Command *command) override;
     States::Type applyActiveOpticForces(ApplyActiveOpticForcesCommand *command) override;
     States::Type clearActiveOpticForces(ClearActiveOpticForcesCommand *command) override;
-    States::Type enableHardpointCorrections(EnableHardpointCorrectionsCommand *command) override;
-    States::Type disableHardpointCorrections(DisableHardpointCorrectionsCommand *command) override;
 
 protected:
     virtual States::Type getLoweringState() = 0;